requirements.txt
anyio==3.6.1
attrs==21.4.0
base58==2.1.1
certifi==2022.5.18.1
ECPy==1.2.5
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
rfc3986==1.5.0
sniffio==1.2.0
tomli==2.0.1
typing-extensions==3.10.0.2
websockets==10.3
xrpl-py @ git+https://github.com/dangell7/xrpl-py.git@076bbb7912072d199033bd50b48e428887164a04
